Speed (#69)
* Update state & county codes only for those that are missing

Recalculating them all takes 2+ hours. This reduces it to only a few
minutes. I think that we should be able to trust new station data to
have the correct values if they are populated.

* better checking to see if file_path is a string

I kept seeing AttributeError’s on the server and TypeError’s on my
machine. So I thought that this check would be safer.